requirements.txt
Django==3.2.7
djangorestframework==3.12.4
psycopg2-binary==2.9.1
django-filter==2.4.0
django-cors-headers==3.8.0
